+++ playbook-website/app/javascript/components/MainSidebar/MenuData/GuidelinesNavItems.ts
