+++ playbook-website/app/javascript/components/MainSidebar/MenuData/SidebarNavItems.ts
